packages/shared_lib/pyproject.toml
[project]
name = "shared-lib"
version = "0.1.0"
description = "Shared database models and utilities"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.12.0",
    "tortoise-orm[asyncpg]>=0.25.3",
    "aiosqlite>=0.17.0",
    "protobuf>=6.33.4",
    "loguru>=0.7.3",
    "pycryptodomex>=3.23.0"
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.76.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

